CaseFilesLH

Handles names and directories for named "Cases" that are typically used to run computations with counterfactuals.

A CaseName consists of a base name, modifiers, and experiment modifiers.

From it, subdirectories are constructed using relative_dir.

CaseName	Directory

:default	"default/base"

:default, [:mod1]	"default/mod1"

:default, [:mod1, :mod2], [:xp1, :xp2]	"default/mod1mod2/experiments/xp1xp2"

CaseFilesLH.CaseName — Type

A CaseName contains the base name, modifiers for the base name, and modifiers for counterfactuals (or experiments).

CaseFilesLH.relative_dir — Function

relative_dir(cn)

Return string with subdirectories where results are stored E.g., "default/smallPrefShocks/experiments/xp1"

CaseFilesLH.base_name — Function

base_name(cn)

Base name.

CaseFilesLH.has_mods — Function

has_mods(cn)

Does case have modifiers?

CaseFilesLH.base_mods — Function

base_mods(cn)

Modifiers of base name.

CaseFilesLH.is_experiment — Function

is_experiment(cn)

Is this Case an experiment?

CaseFilesLH.experiment_mods — Function

experiment_mods(cn)

Experiment modifiers. Empty if not an experiment.