CollegeEntry

This package contains a generic college entry or assignment model. The general idea is to map the values of working and of attending various colleges together with student characteristics, such as test scores, into college entry probabilities.

Once the user has set up an AbstractEntryDecision and an AbstractAdmissionsRule, calling entry_decisions yields entry probabilities and expected utilities for students with given payoffs from attending each college. The results are returned as an AbstractEntryResults object.

For sequential entry mechanisms, students need to be ranked to determine the order in which they get to choose colleges. This is done using AbstractRankings.

Entry is always represented as sequential choice in the rank order of the students. Without capacity constraints, sequential choice and simultaneous choice are the same, so this representation is without loss of generality. Without capacity constraints, the rank order does not matter.

Example:

# There are `nc` colleges
nc = 4;

# Set up admission rules
# Students are ranked by the endowment named `hsGpa`.
# The cutoff for each college that grants admission ranges from 0.0 to 0.8.
admissionS = AdmissionsCutoffSwitches(nc, :hsGpa, 
        collect(range(0.0, 0.8, length = nc)), 0.05);

# Set up entry decision structure
# There are `J` types
J = 200;
# and `nl` locations
nl = 3;
# Total college capacity as multiple of type mass is
totalCapacity = J * nl * 0.4;
switches = make_test_entry_sequ_multiloc(J, nc, nl, totalCapacity);
# Initialize the entry decision
entryS = init_entry_decision(ObjectId(:test), switches, st);

# From model solution, get values of working and of studying
# Provide the ranking of each student according to the indicator that determines the order in which students get to decide (rank_jV).
# Provide the `hsGpa` of each student that is used in admissions.
# The result `er` is an `AbstractEntryResult`
er = entry_decisions(entryS, admissionS,
        vWork_jV, vCollege_jcM, hsGpaPctV, rank_jV);

Notational note: Several functions have suffixes that indicate the dimensions of the objects to be returned. For example, type_mass_jl returns the mass of types by (type, location), whereas type_mass_j returns the total mass of each type across locations.

Notation (Latex symbols) are provided into object constructors as LatexLH.SymbolTable objects. make_test_symbol_table() contains all the objects that need to be defined for all variations of entry scenarios.

Admission rules

An AbstractAdmissionsRule determines the probability that a student may attend a particular college as a function of the students characteristics. Special cases are:

AdmissionsOpen: open admissions.

AdmissionsCutoff: admissions are determined by cutoff values for a single indicator, such as test scores.

Admission rules are constructed from "switches," such as AdmissionsCutoffSwitches using make_admissions.

CollegeEntry.AbstractAdmissionsRule — Type

Abstract type for admissions rules.

CollegeEntry.AdmissionsOpen — Type

Open admissions. Any student may attend any college.

CollegeEntry.AdmissionsCutoff — Type

Admissions are governed by a single indicator, such as a test score. Students can attend colleges for which they qualify in the sense that their indicator exceeds the college's cutoff value. Students may be allowed to attend other colleges with a fixed probability.

CollegeEntry.make_admissions — Function

make_admissions(switches)

Initialize an admission rule from its switches.

Entry decisions

An AbstractEntryDecision provides a structure according to which students make entry decisions, given a set of available colleges. The only concrete type right now is EntryDecision, which can be set up for one location or for multiple with and without capacity constraints.

The idea is to implement an assignment mechanism similar to Hendricks, Herrington, and Schoellman (2021 AEJM). The ingredients are:

There is a set of colleges with fixed qualities and capacities.

A student ranking that determines the order in which students choose colleges. This is provided simply as an ordinal student ranking from outside of the package.

An AbstractAdmissionsRule that determines whether a given student can attend a given college, assuming the college is not full. This could be probabilistic.

The algorithm then proceeds as follows:

Start with all colleges empty.

Loop over students in order of their ranking.

For each student, determine the set of colleges that admit the student (according to the AbstractAdmissionsRule) and that are not full.

Calculate the probability that each student attends each available college.

Record the increment in each college's enrollment. For the next student, colleges that are full are no longer available.

Continue to the next student in the ranking.

Useful generic methods are listed below.

CollegeEntry.AbstractEntryDecision — Type

Abstract type for entry decision protocols.

CollegeEntry.EntryDecision — Type

Sequential entry protocol.

CollegeEntry.init_entry_decision — Function

`init_entry_decision(objId :: ObjectId, switches :: AbstractEntrySwitches)`

Initializes an AbstractEntryDecision from its switches.

CollegeEntry.entry_decisions — Function

entry_decisions(entryS, admissionS, vWork_jV, vCollege_jcM, endowPctV, rank_jV; prefShocks)

Compute entry probabilities and expected values at entry from admission rule and entry decision objects.

It is possible to solve for a subset of types. Only those listed in rank_jV will be solved. But keep in mind that capacity constraints may then not bind.

The type dimension of the EntryResults object matches that of the EntryDecision.

Optional: solve each student's entry decision without preference shocks, but continue to update enrollments with preference shocks. This is useful to diagnose how important preference shocks are.

Note that multiple locations only matter if they are not identical. If all colleges are available in all locations (and not full), the fraction going local (conditional on entry) is only a function of the number of locations and the value of going local. vCollege_jcM and vWork_jV do not matter. This is easy to check analytically.

CollegeEntry.entry_probs — Function

entry_probs(e, vWork_jV, vCollege_jcM, admitV; prefShocks)

Entry probability for a student who is admitted to colleges in admitV. Returns: Entry prob by [type, college], expected value at decision stage by type. This function does not handle the sequential nature of admissions. It is mainly here for unified interface. It defaults to the one step entry protocol. If that does not apply for a protocol, need to define a new method (e.g. two step entry).

CollegeEntry.scale_entry_probs! — Function

scale_entry_probs!(entryProb_jlcM, minEntryProb, maxEntryProb)

Scale entry probabilities by (j,l) -> c to bound away from 0. Bound entry probs by (j,l) away from 1. Does not guarantee particular min values (that is hard to do).

Scale entry probs to bound within min_entry_prob and max_entry_prob.

CollegeEntry.set_local_only_colleges! — Function

set_local_only_colleges!(switches, icV)

Mark which colleges can be attended by locals only.

Extension: Diversity preferences.

Colleges' preferences for diversity may be modeled by imposing quotas on students with certain characteristics. For example, high quality colleges may only admit a certain fraction of students in each income quartile. Once the quota is filled, such colleges are removed from the admissions set for students in those quartiles.

Student Rankings

For sequential admissions protocols, students need to be ranked. All colleges agree on the ranking which determines the order in which students get to choose colleges.

Steps:

Define an object that holds endowment draws.

Extend retrieve_draws and n_draws for this object.

Set up switches that govern the ranking, such as EndowPctRankingSwitches.

Call make_student_ranking to initialize an AbstractRanking.

Call entry_decisions to compute entry probabilities by type and college and expected values at college entry. This accepts the endowment draws object as input and call retrieve_draws to retrieve named endowments.

If student rankings and admissions are supposed to operate on the same object, that object needs to be a single endowment. For example, students can be admitted based on expected abiity given GPA and parental. This is constructed as an endowment draw.

CollegeEntry.AbstractRanking — Type

Abstract student ranking type.

CollegeEntry.AbstractRankingSwitches — Type

Abstract type for switches governing student ranking.

CollegeEntry.EndowPctRanking — Type

Rank students by a linear combination of endowment percentiles.

CollegeEntry.EndowPctRankingSwitches — Type

Switches for linear endowment percentile weights. eNameV contains the endowments used for ranking. It must be possible to call retrieve_draws(EndowmentDraws, eName). wtV are the weights to be used if not calibrated. This omits the first weight, which is fixed at 1. This is empty when there is only one endowment to rank on. doCal determines whether weights are calibrated or fixed.

CollegeEntry.retrieve_draws — Function

retrieve_draws(draws, eName)

Retrieve draws for endowment eName. Needs to be defined for objects passed into ranking functions.

CollegeEntry.n_draws — Function

n_draws(draws)

Number of individuals in endowment draws. Needs to be defined for objects passed into ranking functions.

CollegeEntry.make_student_ranking — Function

Initialize an AbstractRanking object from its switches.

Results

entry_decisions returns an AbstractEntryResults object. This can be queried using a unified interface, even though different entry protocols produce information in different ways. For example, only some protocols have locations.

CollegeEntry.AbstractEntryResults — Type

Abstract type for results from entry decisions.

CollegeEntry.EntryResults — Type

Entry results for multiple locations. Records:

fracEnter_jlcM: Fraction of students in (j,l) who attend college c

fracLocal_jlcM: The same, but for the local college c (if any).

fracEnterBest_jlcM: Fraction of students in (j,l) who attend college c and for who this is the best available college.

enroll_clM: total enrollment at college (c, l).

enrollLocal_clM: local enrollment at college (c, l).

enrollBest_clM: enrollment of students for who this is the best available college.

eVal_jlM: expected value of students in (j,l) at the start of the process.

CollegeEntry.n_locations — Function

n_locations(e)

Number of locations.

CollegeEntry.n_colleges — Function

n_colleges(e)

Number of colleges.

CollegeEntry.n_types — Function

n_types(e)

Number of student types.

CollegeEntry.capacities — Function

College capacities. Set to an arbitrary large number for entry mechanisms where capacities do not matter.

CollegeEntry.enrollment_cl — Function

College enrollments. By location (if any).

CollegeEntry.enrollment_c — Function

Total enrollment by college; across all locations.

CollegeEntry.entry_probs_jlc — Function

Entry probabilities by college for each type, location.

CollegeEntry.entry_probs_jc — Function

entry_probs_jc(e)
entry_probs_jc(e, univ)

Fraction of each type entering each college; across all locations.

CollegeEntry.entry_probs_j — Function

entry_probs_j(er)
entry_probs_j(er, univ)

Entry probabilities by type, across all colleges.

CollegeEntry.expected_values_jl — Function

Expected value at college entry stage, by type.

CollegeEntry.frac_local — Function

Fraction of students who attend local colleges.

CollegeEntry.frac_local_j — Function

Fraction of students of each type who attend local colleges.

CollegeEntry.frac_local_c — Function

Fraction of students at each college who are local. Across all locations.