CollegeStratCollege

This package collects the college related code for the CollegeStrat project.

A college consists of

a dropout rule

a graduation rule

a course grid

a human capital production function

a human capital shock process

a learning specification that determines how learning is split between the two types of human capital

a tuition function

choice sets for study and work times

and some other details

Dropout rule

The dropout rule determines when a student must drop out of college.

CollegeStratCollege.DropoutRule — Type

Dropout rule for one college. Constructed from the DropoutRuleSet.

CollegeStratCollege.DropoutRuleSet — Type

Specifies dropout rules for a set colleges.

CollegeStratCollege.DropoutRuleSwitches — Type

Switches used to construct a DropoutRuleSet.

CollegeStratCollege.DropoutRuleSimple — Type

Simple DroputRule.

CollegeStratCollege.DropoutRuleSetSimple — Type

Simple DropoutRuleSet where students must drop out after a fixed number of years in college.

CollegeStratCollege.DropoutRuleSwitchesSimple — Type

Switches for constructing a simple DropoutRuleSet.

CollegeStratCollege.make_dropout_set — Function

make_dropout_set(objId, switches)

Construct a DropoutRuleSet from its switches.

CollegeStratCollege.make_dropout_rule — Function

make_dropout_rule(ds, iCollege)

Construct the DropoutRule for one college.

CollegeStratCollege.drop_prob — Function

drop_prob(dr, t, h, n)

Dropout probability at end of t. Inputs are h, n at END of t.

CollegeStratCollege.drop_prob_grid — Function

drop_prob_grid(g, t, hV, nTakenV)

Dropout probability on an [h, n] grid; at end of t.

Graduation Rule

The graduation rule gives the probability that a student may graduate at the end of each year in college.

Course grid

The course grid determines the number of courses a student make take in each period.

Human capital production function

This determines how much students learn in college.