CommonLH

This package collects general purpose code that does not "fit" into other packages with more well defined topics.

Checking properties

CommonLH.check_float — Function

check_float(x; lb, ub)

Check a scalar: should be real, finite, not nan, in optional bounds. Real is implied by being AbstractFloat.

CommonLH.check_float_array — Function

check_float_array(m, lb, ub)

Check that a numeric array is finite, not NaN, and inside bounds.

Array operations

Missing docstring.

Missing docstring for bracket_array!. Check Documenter's build log for details.

CommonLH.scale_array! — Function

scale_array!(x, d, totalV)

Scale each "row" (along dimension d) of array x to match the totals given in totalV. The result satisfies

sum(selectdim(x, d, j)) == totalV[j]

Missing docstring.

Missing docstring for all_at_least. Check Documenter's build log for details.

Missing docstring.

Missing docstring for all_at_most. Check Documenter's build log for details.

Missing docstring.

Missing docstring for all_greater. Check Documenter's build log for details.

Missing docstring.

Missing docstring for all_at_most. Check Documenter's build log for details.

Missing docstring.

Missing docstring for any_greater. Check Documenter's build log for details.

Missing docstring.

Missing docstring for any_less. Check Documenter's build log for details.

Missing docstring.

Missing docstring for any_nan. Check Documenter's build log for details.

Vector operations

Missing docstring.

Missing docstring for bisecting_indices. Check Documenter's build log for details.

CommonLH.find_indices — Function

find_indices(valueV, gridV; notFoundError)

For each value in valueV (a weakly increasing Vector), find the corresponding indices in gridV (also an increasing Vector{Integer}). If no match is found, set indices to 0. Returns a Vector even if only one value is given.

Example

julia> find_indices([2, 4], -2 : 2 : 10) == [3, 4]
true

Missing docstring.

Missing docstring for find_index. Check Documenter's build log for details.

Probabilities

The following operate on matrices that represent probabilities which sum to 1.

CommonLH.validate_prob_matrix — Function

validate_prob_matrix(probM)

Validate a probability matrix. Elements are expected to sum to 1.

CommonLH.validate_prob_vector — Function

validate_prob_vector(probV)

Validate a probability vector. Elements are expected to sum to 1.

CommonLH.prob_j — Function

Conditional probabilities of rows or columns given a probability matrix.

CommonLH.prob_k — Function

Conditional probabilities of rows or columns given a probability matrix.

CommonLH.prob_j_k — Function

prob_j_k(m, j, k)

Conditional probability (j | k).

CommonLH.prob_k_j — Function

prob_k_j(m, k, j)

Conditional probability (k | j).

CommonLH.ev_given_j — Function

ev_given_j(x, prM, j)

Compute expected value of x, given row index j. Based on matrix of values xM[j,k] and matrix of probabilities (sum to 1) prM[j, k].

CommonLH.ev_given_k — Function

ev_given_k(x, prM, k)

Compute expected value of x, given column index k. Based on matrix of values xM[j,k] and matrix of probabilities (sum to 1) prM[j, k].

CommonLH.scale_prob_array! — Function

scale_prob_array!(m; maxSum)

Given an array of probabilities: ensure that all are in [0, 1]. Error if bounds violation larger than rounding errors. Make sure that sum does not exceed an upper bound.

Displaying objects

CommonLH.MultiIO — Type

Object that holds multiple IO streams, so that a single print statement can be directed to all of them. Main use case: selectively write to optimization log and stdout. This must be a subtype of IO for dispatch to work correctly.

Based on https://discourse.julialang.org/t/issues-with-println-buffering-output-when-redirecting-stdout-to-a-txt-file/23738/13

CommonLH.show_text_file — Function

show_text_file(filePath; io)

Show a text file on the screen.

CommonLH.show_string_vector — Function

show_string_vector(sV)
show_string_vector(sV, width)

Display string vector on fixed with screen.