DeviationsLH

A package to keep track of deviations between data moments and model moments. This is meant to be used when calibrating economic models.

Data moments and deviations

The AbstractDevation object is designed to keep track of target moments for the calibration. It also stores the corresponding model moments and can therefore compute and display measures of model fit.

There are several types of AbstractDeviations:

Deviation is the default type. It holds Arrays of AbstractFloatss of any dimension.

ScalarDeviation holds scalar moments.

RegressionDeviation handles the case where the target moments are represented by regression coefficients and their standard errors.

BoundsDeviations are zero until the model values get out of bounds. These are useful for preventing numerical optimizers from trying "crazy" parameter values.

scalar_dev computes a scalar deviation from model and data values (weighted).

DeviationsLH.AbstractDeviation — Type

AbstractDeviation

Abstract type for Deviation objects. A Deviation object implements:

* `set_model_values`
* `scalar_dev`

Contains:

* `stdV`: std errors of data. Defaults to []. Used when showing deviations.
* `wtV`: weights within a deviation (e.g. 1/std error)
* `scalarWt`: used by `DevVector` to weight each `scalar_dev`
* `showFct`: function that takes an `AbstractDeviation` as input and produces a model/data comparison
* `showPath`: file name where `showFct` files its output. `stdout` is used if empty.

DeviationsLH.Deviation — Type

Deviation

Holds numeric arrays. The default for deviations.

Contains model values, data values, data std deviations (for display only), weights (for constructing scalar deviation).

Can specify that only modelV[idxV...] is to be compared with dataV.

DeviationsLH.ScalarDeviation — Type

ScalarDeviation

Here the wtV field is intended to hold 1 / std error of the moment. The scalar deviation is always the absolute deviation. For scalars, the exponent in the norm makes no difference.

DeviationsLH.RegressionDeviation — Type

RegressionDeviation

Holds model and data in the form of RegressionTable objects

DeviationsLH.BoundsDeviation — Type

Bounds deviation. Returns zero scalar deviation until model values get out of bounds.

DeviationsLH.scalar_dev — Function

Scalar deviation from one Deviation object.

Optionally includes scalarWt factor. Uses a weighted norm. The exponent is determined by the Deviation's normP property.

Scalar deviation from one Deviation object.

Optionally includes scalarWt factor.

Scalar deviation from one Deviation object.

Optionally includes scalarWt factor.

ChangeTable

This is an object that can keep track of which model parameters affect which deviations.

The user initializes the table with the constructor. Then they solve the model perturbing one parameter at a time. ChangeTable produces a formatted table that shows the effect of each parameter on each deviation.

DeviationsLH.ChangeTable — Type

Table that shows how each parameter affects each deviation. Rows are deviations. Columns are parameters.