DistributionsLH

This package makes ModelObjects for random variables drawn from select distributions.

For the most part, this just wraps distributions from Distributions.jl into ModelObjects from ModelParams so that the parameters can be calibrated conveniently and so that fixed properties can be set using switches objects.

There are also extensions to the distributions in Distributions.jl. For example, the Beta distribution is defined over an arbitrary range rather than [0, 1].

Generic interface

Each distribution implements the following

DistributionsLH.init_distribution — Function

init_distribution(objId, switches)

Initialize a distribution from its switches.

DistributionsLH.draw — Function

draw(distrib, nDims, rng)
draw(distrib, rng)

Draw from the distribution.

DistributionsLH.scale_normal_draws — Function

scale_normal_draws(b, inM)

Converts N(0,1) draws into the specified distribution.

DistributionsLH.quantiles — Function

quantiles(distrib, pctV)

Quantiles.

DistributionsLH.min_value — Function

min_value(distrib)

Minimum value (if any).

DistributionsLH.max_value — Function

max_value(distrib)

Maximum value (if any).

DistributionsLH.check_draws — Function

check_draws(distrib, drawM)

Check draws against bounds (if any) and Inf values.

Uniform distribution

DistributionsLH.AbstractUniformSwitches — Type

Abstract type for switches governing how Uniform distributions are parameterized.

DistributionsLH.AbstractUniform — Type

Abstract Uniform type.

DistributionsLH.Uniform — Type

Unform distribution on interval xMin + [0, xRange].

DistributionsLH.UniformFixedBounds — Function

UniformFixedBounds(objId, lb, ub)

Uniform distribution with fixed bounds (not calibrated).

DistributionsLH.UniformCenteredSwitches — Type

Switches for Uniform distribution that is centered around a mean.

DistributionsLH.UniformCentered — Type

Unform distribution on interval xMean +/- 0.5 * xRange.

DistributionsLH.init_uniform — Function

init_uniform(objId, switches)

Construct a Uniform distribution from its switches.

Beta distribution

DistributionsLH.BetaSwitches — Type

Switches for default Beta distribution object.

DistributionsLH.Beta — Type

Beta marginal distribution. Characterized by lower bound, upper bound, and Beta parameters alpha and beta. Cannot store the Beta distribution in the object because its parameters are not known.

DistributionsLH.init_beta — Function

init_beta(objId, switches)

Construct a Uniform distribution from its switches.

Multivariate Normal Distribution

The main purpose is to be able to draw random variables from the "weight matrix" which is basically a lower triangular decomposition of the covariance matrix, but with ones on the diagonal. The draws are then scaled to match the target means and standard deviations of the marginals.

Also computes conditional distributions.

DistributionsLH.MvNormalLHSwitches — Type

Switches used to construct Multivariate Normal object.

DistributionsLH.MvNormalLH — Type

Multivariation Normal object.

This is a ModelObject but currently does not support calibrating parameters.

DistributionsLH.cov_matrix — Function

Make cov matrix from weight matrix.

The weight matrix is lower triangular with ones on the diagonal.

DistributionsLH.check_cov_matrix — Function

Check that a covariance matrix is valid for a given MvNormalLH object.

DistributionsLH.draw_from_weights — Function

draw_from_weights(m, wtM, nObs, rng)

Draw random variables given the weight matrix. Returns Matrix by observation, variable.

DistributionsLH.check_weight_matrix — Function

check_weight_matrix(m, wtM)

Check that weight matrix is valid.

DistributionsLH.conditional_distrib — Function

Compute conditional distribution of a set of variables, given the others For Multiple sets of conditioning observations

Arguments

covM covariance matrix

idx2V indices of variables on which we condition

value2M[observation, variable] their values

Outputs

condMeanM[observation, variable]

condStdV(variable) conditional means and std of each variable, given all others for those not in idx2V

condCovM(variable, variable) conditional covariance; for those not in idx2

DistributionsLH.cond_mean_weights — Function

cond_mean_weights(m, idx2V, covM)

Weights of conditioning variables for conditional means: sigma12/sigma22. Such that

`conditional mean = mu1 + weights * (value2 - mu2)`

Outputs

wtM n1 x n2 matrix of weights each row gives the weights for a "dependent" variable