EconLH

Discrete choice

This solves discrete choice problems subject to i.i.d. Gumbel preference shocks. An ExtremeValueDecision object can be constructed. Or extreme_value_decision can be called directly.

EconLH.ExtremeValueDecision — Type

Solve a decision problem with type I extreme value shocks

Arguments

value_ixM values of alternatives rows are types; columns are alternatives works with a single type

prefScale scale of type I extreme value shocks

EconLH.demeaned — Function

Use demeaned Gumbel preference shocks?

EconLH.pref_scale — Function

Returns the scale parameter of the Gumbel preference shocks.

EconLH.extreme_value_decision — Function

extreme_value_decision(d, value_ixM)

Decision: Multiple alternatives. Always returns a vector for eVal_iV (even with one agent).

OUT prob_ixM probability of choosing each option eVal_iV expected value for each type

TEST by simulation in extreme_value_decision_test

EconLH.extreme_value_decision_one — Function

extreme_value_decision_one(valueV, prefScale; demeaned)

Extreme value decision for one individual.

This is probably not efficient. See efficient implementation for logsumexp in StatsFuns.jl. But here we also need the scaled exp(valueV) directly for the probabilities.

EconLH.draw_gumbel_shocks — Function

draw_gumbel_shocks(rng, prefScale, sizeV; demeaned)

Draw Gumbel shocks, optionally demeaned. Returns vector, even if sizeV == 1.