EndowmentsLH

The purpose of this package is to provide a uniform interface for storing, accessing, and validating endowment draws in heterogeneous agent economic models.

Endowment draws are collected in the EndowmentDraws object. These draws can be scalars or custom, user defined data types.

Each Endowment contains information about its marginal distribution, encoded as an AbstractMarginal. This allows automatic checking of types and bounds. The following marginal distributions are predefined:

UniformMarginal

NormalMarginal

BetaMarginal

PercentileMarginal

BoundedMarginal

UnboundedMarginal

All of these handle scalar Float or Integer draws, which constitute the majority of the draws in the kinds of models that I work with. The user can define additional marginal distributions (see below).

Each marginal comes with a validate_draws function that validates that endowment draws are valid for this marginal.

Bounded marginals define lb and ub. Where known, quantiles can be looked up with marginal_quantile.

EndowmentsLH.EndowmentDraws — Type

EndowmentDraws()

Constructor for EndowmentDraws.

EndowmentsLH.Endowment — Type

Meta information about one endowment. Contains a name, label, and its marginal distribution.

Example

e = Endowment(:afqt, "Test score", PercentileMarginal{Float64}())

EndowmentsLH.AbstractMarginal — Type

Abstract type for marginal distributions of endowments.

EndowmentsLH.UniformMarginal — Type

Uniform marginal distribution over fixed bounds.

EndowmentsLH.NormalMarginal — Type

Normal(mean, std) marginal. Should store the Normal object +++

EndowmentsLH.BetaMarginal — Type

Beta marginal distribution. Characterized by lower bound, upper bound, and Beta parameters alpha and beta.

EndowmentsLH.PercentileMarginal — Type

Marginal distribution that is Uniform[0, 1].

EndowmentsLH.UnboundedMarginal — Type

Marginal with unknown distribution and no bounds.

EndowmentsLH.validate_draws — Function

Validate draws from a given marginal distribution. Mainly bounds checking and avoiding Inf and NaN.

Missing docstring.

Missing docstring for lb. Check Documenter's build log for details.

Missing docstring.

Missing docstring for ub. Check Documenter's build log for details.

EndowmentsLH.marginal_quantile — Function

marginal_quantile(u, pct)

Quantiles for a marginal distribution. Inputs are percentiles in [0, 1]. Not defined for all distributions. Fallback returns nothing.

Typical Flow

A collection of EndowmentDraws is typically set up by:

Initializing an empty object: ed = EndowmentDraws().

Defining one endowment at a time: endow2 = Endowment(:endow2, "Endow2", PercentileMarginal{Float32}()).

Drawing the endowment (model specific).

Adding the endowment to the object: add_draws!(ed, endow2, drawV).

Validating the draws with validate_draws.

Endowment draws may be retrieved one endowment at a time using get_draws or all endowments for one individual in one go using type_endowments.

EndowmentsLH.add_draws! — Function

add_draws!(e, endow, dV)

Add draws for one variable to existing EndowDraws.

EndowmentsLH.get_draws — Function

get_draws(e, eName)

Retrieve draws for one endowment; for all individuals.

EndowmentsLH.type_endowments — Function

type_endowments(e, j)

Retrieve endowment draws for one type as a DataFrame row. One can splat this as a tuple into a function:

foo(type_endowments(e, j)...) is the same as foo(x, y)