LatexLH

Writing Beamer Slides

LatexLH.write_figure_slides — Function

write_figure_slides(io, slideV; baseDir)

Make a file that contains a set of slides. Each slide is just a title with one figure. Input is a vector of tuples. Each contains a title and a file path for the figure file. Optionally, preprend a common base directory to each path.

LatexLH.figure_slide — Function

figure_slide(title, fPath)

Latex lines for slide with a single figure.

Writing Latex Tables

LatexLH.CellColor — Type

Defines Latex cell colors. Stores name and intensity of a color.

Example

CellColor("blue", 75)

LatexLH.Cell — Type

Single or multicolumn cell. Stores text, width, alignment, color.

LatexLH.Table — Type

Table

Holds a Latex table, constructed as a vector of rows.

Flow

Constructor

Make and add rows (some are just provided by user as strings)

Write table

LatexLH.color_string — Function

color_string(c)

Returns a string of the form "\cellcolor{blue!75}".

LatexLH.cell_string — Function

cell_string(c)

Returns a Latex string that makes a cell. Either just the (colored) contents or a multicolumn command.

LatexLH.nrows — Function

nrows(tb)

Number of table rows.

LatexLH.add_row! — Function

add_row!(tb, rowStr)

Add a row to a table.

LatexLH.make_row — Function

make_row(tb, cellV)

Make a row from a vector of cells

LatexLH.write_table — Function

Write table to file.

SymbolTables

Intended for looking up names and descriptions of model symbols in code. E.g., for correctly naming objects in tables. Also for writing preamble files with model notation.

LatexLH.SymbolInfo — Type

Information for one symbol: name, latex symbol, description, group.

For printing: \ needs to be escaped as \\ in latex.

LatexLH.SymbolTable — Type

SymbolTable()

Constructor for empty SymbolTable.

SymbolTable(m)

Constructor from String array with names, latex symbols, descriptions, groups.

Example

SymbolTable(["capShare"  "\alpha"  "Capital share"  "Technology";
    "tfp"  "A"  "Total factor productivity"  "Technology"]);

LatexLH.add_symbol! — Function

add_symbol!(s, si; replaceExisting)

Add a new symbol. Overwrite existing if replaceExisting == true.

LatexLH.has_symbol — Function

has_symbol(s, name)

Does symbol exist?

LatexLH.description — Function

description(s, name)

Return description of a symbol.

LatexLH.latex — Function

latex(s, name)

Latex symbol for a symbol.

LatexLH.group — Function

group(s, name)

Group for a symbol.