ModelObjectsLH

This package defines the abstract ModelObject type and the ObjectId concept.

The idea is to provide a way of keeping track of where an object is located inside a model that consists of nested ModelObjects. This is useful for finding objects by ObjectId and for keeping track of where model parameters "belong."

ModelObject

The abstract type ModelObject determines which objects the methods of ModelParams work on. Objects that are not subtypes of ModelObject are ignored.

A ModelObject typically contains other ModelObjects. Each contains (potentially) calibrated parameters.

ModelObjectsLH.ModelObject — Type

ModelObject

Abstract model object Must have field objId :: ObjectId that uniquely identifies it May contain a ParamVector, but need not.

Child objects may be vectors. Then the vector must have a fixed element type that is a subtype of ModelObject

ObjectId

Each ModelObject has a unique ObjectId. It identifies where each object is located in the model hierarchy. Example: ":model > :firm > :technology" would be (the string representation of) and ObjectId.

The ObjectId keeps track of the parent object and of an index. The index is used when a vector of objects is created. For example, if we have several household types, their ObjectIds might be :hh[1], :hh[2], etc.

ObjectIds are automatically unique. Even if and object named :Foo occurs in different child objects of a ModelObject, the ObjectIds will be unique.

Using an ObjectId, the corresponding object can be located inside the parent object without ambiguity.

make_string produces a String representation of an ObjectId. This is useful for writing ObjectIds to text based data structures, such as JSON files. make_single_id does the reverse: it reconstructs the ObjectId from the string generated by make_string. This is used for reading ObjectId from text based files.

ObjectIds can be constructed in several ways. make_object_id creates an ObjectId from its string representation.

ModelObjectsLH.ObjectId — Type

ObjectId

Complete, unique ID of a ModelObject

Contains own id and a vector of parent ids, so one knows exactly where the object is placed in the model tree.

Missing docstring.

Missing docstring for get_object_id. Check Documenter's build log for details.

ModelObjectsLH.make_string — Function

make_string(id)

Make a string from a SingleId. Such as "x[2, 1]".

make_string(id)

Make string from ObjectId. Such as "p > q > r[4, 2]".

ModelObjectsLH.make_child_id — Function

make_child_id(obj, name)
make_child_id(obj, name, index)

Make child ObjectId from parent ObjectId.

ModelObjectsLH.make_single_id — Function

make_single_id(s)

The inverse of make_string.

ModelObjectsLH.make_object_id — Function

The inverse of make_string.

ModelObjectsLH.own_name — Function

own_name(oId)

Return object's own name as Symbol.