ModelParams

The purpose of ModelParams is to offer an automated workflow for handling model objects and their potentially calibrated or estimated parameters.

Background

The target audience consists mainly of economists who calibrate or estimate structural models.

One may think of a model as a collection of model objects, such as utility functions, production functions, etc. Model objects may contain nested model objects. For example, a household model object may contain a utility function.

Typically, a researcher will calibrate many model versions that differ in:

functional forms (e.g., log utility versus CARA)

which parameters are fixed versus calibrated

switches (e.g., does the model have preference shocks? How many types of households are there?)

The challenge is then to keep track of:

which fixed parameters need to be set?

which parameters need to be calibrated?

where to put the calibrated parameter values in the model?

In addition, one needs to keep track of which data moments are used in the calibration and what the associated model moments are.

All of this is automated in ModelParams.

To see in action how this works, see the SampleModel repo in my github account.

Parameters: Param

Model parameters that are either calibrated or fixed are defined in Param objects. These encode default values (used when not calibrated), bounds, Latex symbols, and descriptions.

Vectors that are increasing are encoded in an IncreasingVector.

ModelParams.IncreasingVector — Type

IncreasingVector

Encodes an increasing vector of fixed length. Its values are calibrated.

Base.values — Function

Retrieve values of an IncreasingVector.

Returns all values of a BoundedVector.

values(iv, idx)

Returns a subset of the values of a BoundedVector.

BoundedVector encodes a vector that is monotone and bounded.

ModelParams.BoundedVector — Type

Increasing or decreasing vector with bounds. The special case where the vector is of length 1 is supported.

A BoundedVector is typically constructed with an empty ParamVector. Then set_pvector! is used to initialize the ParamVector. The ParamVector contains a single entry which must be named :dxV. It sets the values for the eponymous BoundedVector field. The dxV are typically in [0, 1]. They represent the increments in the vector.

ModelParams.set_pvector! — Function

set_pvector!(iv; description, symbol, isCalibrated)

Initialize the ParamVector. Requires dxV to be set. Note that bounds on dxV must be between 0 and 1. This is called after the BoundedVector has been constructed with an empty ParamVector but with the values of the field dxV set to reasonable defaults.

Example

dxV = [0.3, 0.2, 0.8];
b = BoundedVector(objId, ParamVector(objId), true, 1.0, 2.0, dxV);
set_pvector!(b; description = "Gradient", symbol = "g(x)");

ModelParams.set_default_value! — Function

set_default_value!(iv, valueV)

Set the default values of a BoundedVector.

ModelParams.fix_values! — Function

Switches calibration toggle off. Sets values and default values everywhere. The end result is a BoundedVector with fixed (not calibrated) increments that result in values of valueV.

Parameter Vectors: ParamVector

The second key concept is a ParamVector. It collects all potentially calibrated model parameters, represented as Param objects.

For convenience and performance, a ModelObject has a field for each Param. The field values should never be written to "by hand." Functions are provided that sync fixed and calibrated parameters from the ParamVector.

To enable unambiguous matching of ParamVectors to child objects, each ParamVector contains the object's ObjectId.

ModelParams.ParamVector — Type

ParamVector

Vector containing all of a model's potentially calibrated parameters. Parameters contain values, not just default values They are kept in sync with values in object

Intended workflow: See SampleModel Create a model object with parameters as fields Otherwise the code gets too cumbersome Constructor initializes ParamVector with defaults (or user inputs) During calibration Each object generates a Dict of calibrated parameters Make this into a vector of Floats that can be passed to the optimizer. Optimization algorithm changes the floats Make floats back into Dict Copy back into model objects

Going from a vector of Dicts to a vector of Floats and back: make_guess set_params_from_guess! These are called on the top level model object

ToDo: Make the process of going from model -> vector and vice versa more robust.

Currently, the user has to ensure that the ordering of ParamVectors and model
objects never changes.

ModelParams.Param — Type

Param

Holds information about one potentially calibrated parameter (array). Default value must always be set. Determines size of inputs. Everything else can be either empty or must have the same size.

Useful access routines for ParamVectors include:

ModelParams.retrieve — Function

retrieve

Returns a named parameter and its index in the ParamVector. First occurrence. Returns 0 if not found.

Base.append! — Function

append!

Append a Param to a ParamVector

ModelParams.remove! — Function

remove!(pvec, pName)

Remove the parameter names pName from pvec.

ModelParams.replace! — Function

replace!(pvec, p)

Replace a parameter with a new parameter p.

ModelParams.change_calibration_status! — Function

change_calibration_status!(pvec, pName, doCal)

Set whether or not a parameter is calibrated.

ModelParams.change_value! — Function

change_value!(pvec, pName, newValue)

Change the value of parameter pName.

change_value!(x, oName, pName, newValue)

Change value of a field in a ModelObject and its ParamVector.

Vectors of ParamVector

These are used when all the ParamVectors from a ModelObject are collected.

ModelParams.find_pvector — Function

find_pvector(pvv, objId)

Find a ParamVector for a given ObjectId. Returns index and the ParamVector. Or 0 and nothing if not found.

Calibrating a model

The workflow is implemented and tested in the SampleModel repo.

Initialize a model object, including its child objects. E.g., SampleModel.Model.

Write the code and solves the model and computes any desired statistics. This part is not affected by anything related to ModelParams because the potentially calibrated parameters are stored inside each object and synced with the ParamVectors.

Write a deviation function that accepts a vector of calibrated parameters as an input. Call set_params_from_guess! to copy the parameter values from the vector into the various model objects.

In the calibration function, call make_guess to make a vector of parameter values from the parameter values in all of the model objects.

Report the calibrated and fixed parameters using report_params

ModelParams.set_params_from_guess! — Function

set_params_from_guess!(m, guess)

Make vector of guesses into model parameters. For object and children. This changes the values in m and in its pvector.

ModelParams.make_guess — Function

make_guess(m)

Make vector of parameters and bounds for an object Including nested objects

ModelParams.report_params — Function

report_params(pvec, isCalibrated; io, closeToBounds)

Reports calibrated (or fixed) parameters for one ParamVector as a PrettyTable.

report_params(o, isCalibrated; io, closeToBounds)

Report all parameters by calibration status. For all ModelObjects contained in o.

Intended for reporting at the end (or during) a calibration run. Not formatted for inclusion in papers.

Each table row looks like: "Description (name): value"

ModelParams.param_tables — Function

param_tables(o, isCalibrated)

Generate a set of ParamTables. One table per model object. These are stored in a Dict with ObjectId's as keys.

The purpose is to make it easy to generate nicely formatted parameter tables that are grouped in a sensible way.