ModelStatisticsLH

A package that offers containers for storing model generated statistics by group.

VarInfo

Contains information about a variable:

name

description

data type and size

optional additional meta info stored as key value pairs

Example:

vi = VarInfo(:x, "This is x", Vector{Int}; size = (3,), lb = 2);
eltype(vi) == Vector{Int};
size(vi) == (3,);

It is possible to add arbitrary user defined objects as long as they support Base.size.

ModelStatisticsLH.VarInfo — Type

Holds meta information about one variable. options holds predefined options and user defined options. Predefined options are:

size (required)

lb, ub

lbs, ubs

dimSums

grandSum

These can be used to check a data matrix.

Example

vi = VarInfo{Vector{Int}}(:x, "x stores a Vector of Int"; size = (3,), lb = 0);
eltype(vi) == Int;

ModelStatisticsLH.get_options — Function

get_options(vi, oNames)

Get multiple options. nothing if not found.

Missing docstring.

Missing docstring for get_options. Check Documenter's build log for details.

Missing docstring.

Missing docstring for has_option. Check Documenter's build log for details.

ModelStatisticsLH.var_name — Function

Returns name of the variable.

Missing docstring.

Missing docstring for description. Check Documenter's build log for details.

ModelStatisticsLH.bounds — Function

Returns the bounds of the variable. These are nothing if not set.

Missing docstring.

Missing docstring for init_values. Check Documenter's build log for details.

ModelStatisticsLH.check_var — Function

Check an array against the constraints implied by a ArrayInfo.

Arguments

silent: governs whether @warn messages are shown.

ModelStats

Holds statistics and meta information.

Values can be accessed and set using dot notation:

g.x == get_values(g, :x);
z = g.x[1:2, 3, 4];
g.x = [1,2,3];
g.x[1:2, 3:5] = rand(2,3);
g.x[1:2, 3:5] .= 0.5;

ModelStatisticsLH.ModelStats — Type

Container that holds model statistics (arrays or scalars) and meta information. For each variable, a VarInfo object holds info needed to display and check the data.

The T parameter is not used anywhere, but it enables the user to construct different parametric ModelStats{T} and dispatch on those.

ModelStatisticsLH.get_meta — Function

Retrieve meta info (not for a variable).

ModelStatisticsLH.has_variable — Function

Do GroupStats contain variable vName?

ModelStatisticsLH.n_vars — Function

Number of variables.

ModelStatisticsLH.var_names — Function

Return all variable names.

ModelStatisticsLH.var_meta — Function

var_meta(g, vName)

Return meta info for a variable. Errors if not found.

Missing docstring.

Missing docstring for validate_variable. Check Documenter's build log for details.

ModelStatisticsLH.validate_stats — Function

Check that

all data are inside bounds

Missing docstring.

Missing docstring for data_table. Check Documenter's build log for details.

ModelStatisticsLH.add_variable! — Function

add_variable!(g, vInfo, newValues)

Add a variable.

Arguments

vInfo: Can be a VarInfo or a Tuple of values that can be fed into the VarInfo constructor.

Example

add_variable!(g, (:vName, 0.0, 1.0), [1,2,3]);

ModelStatisticsLH.delete_variable! — Function

delete_variable!(g, vName)

Delete a variable. Errors if it does not exist.

ModelStatisticsLH.get_values — Function

Return values for a variable. Errors if not found, unless defaultValue is provided.

ModelStatisticsLH.set_values! — Function

Set values for a variable. Errors if variable does not exist. Optional: validates values against meta info.

Example

set_values!(g, :x, [1,2,3]);
# This can also be called with dot notation
g.x = [1,2,3];

StatsCollection

Container that holds several ModelStats objects.

Stats can be accessed by providing "groups" which are the T in the ModelStats{T} objects. Or they can be accessed as if the StatsCollection were a flat collection.

The user defines statname_from_name and groups_from_name to translate a variable name of the form (say) :workTime_gp into the groups (:gpa, :parental) and the variable :workTime. Then retrieving :workTime_gp is the same as retrieving the ModelStats{(:gpa, :parental)} object's :workTime variable.

Example

sc = StatsCollection{:x}();
# Add the statistics, by group
grp = (:gpa, :parental);
ms = ModelStats{grp}(...);
add_mstats!(sc, ms);
ms == get_mstats(sc, grp);
# Different ways of retrieving stats
x1 = get_stats(sc, grp, :workTime);
x1 == get_stats(sc, :workTime_gp);
x1 == sc.workTime_gp;
# getindex is overloaded
sc.workTime_gp[3] == 0.5;
# setindex is overloaded
sc.workTime_gp[3] = 0.5;
# setproperty is overloaded
sc.workTime_gp = newValues;

ModelStatisticsLH.StatsCollection — Type

Container for a collection of ModelStats. May also contain nested StatsCollections.

ModelStatisticsLH.add_mstats! — Function

add_mstats!(m, s)

Add ModelStats or a nested StatsCollection.

ModelStatisticsLH.delete_mstats! — Function

delete_mstats!(m, msName)

Delete ModelStats.

ModelStatisticsLH.replace_mstats! — Function

Replace ModelStats. Add it if it does not exist.

ModelStatisticsLH.get_groups — Function

List of all the groups. That is, m contains ModelStats{grp} for each grp in get_groups(m).

ModelStatisticsLH.has_mstats — Function

Does the ModelStats{groups} object exist?

ModelStatisticsLH.get_mstats — Function

Retrieve the ModelStats{groups} object. Errors if groups not found.

ModelStatisticsLH.get_stats — Function

get_stats(m, groups, statName)

Retrieve a statistic. Groups are specified.

get_stats(m, sg)

Retrieve a statistic without providing groups.

Example

get_stats(m, :workTime_gp) == get_stats(m, (:gpa, :parental), :workTime)
# Equivalently
m.workTime_gp == get_stats(m, :workTime_gp)
m.workTime_gp[2,3] == get_stats(m, :workTime_gp)[2,3]

ModelStatisticsLH.set_stats! — Function

set_stats!(m, groups, statName, x)

Set values for a statistic.

Example

set_stats!(m, (:gpa, :parental), [1,2,3])
set_stats!(m, :workTime_gp, [1,2,3])
# Also with indices
m.workTime_gp[2,3] = x

ModelStatisticsLH.groups_from_name — Function

groups_from_name(m, statName)

Mapping of name of a statistic into groups. User defines this for each parametric StatsCollection{T}.

Example:

groups_from_name(m, :workTime_gp) == (:gpa, :parental)

ModelStatisticsLH.statname_from_name — Function

statname_from_name(m, statName)

Mapping of name of a statistic into its name without the group info. User defines this for each parametric StatsCollection{T}.

Example:

groups_from_name(m, :workTime_gp) == :workTime