StructLH Documentation

Code for manipulating structs.

Helpers

StructLH.@common_fields — Macro

Macro to set common fields in structs. See https://discourse.julialang.org/t/julia-learning-macros-metaprogramming/45753/3

Example

@common_fields set1 begin
    x :: Int
    y :: Float64
end

struct Foo
    @set1
    z
end

StructLH.retrieve_property — Function

retrieve_property(o, pName)

Retrieve a property from an object or its child objects. If the property is not unique, returns the first match and then stops. Returns nothing if not found.

StructLH.retrieve_child_property — Function

retrieve_child_property(o, childName, pName)

Retrieve a property in a child object of a given name. Does not retrieve property from object itself. This helps when properties are not unique.

StructLH.has_property — Function

has_property(o, pName)

Does object or a child object have a property?

Merging objects

StructLH.merge_object_arrays! — Function

merge_object_arrays!

Merge all arrays (and vectors) from object oSource into the corresponding arrays in another object oTg (at given index values idxV). If target does not have corresponding field: behavior is governed by skipMissingFields.

Arguments

idxV Vector{Integer} or other iterable with integer results. Indexes the first dimension of each object to be copied. The assignment is: oSource.x[i1,:] => oTg.x[idxV[i1],:].

StructLH.reduce_object_vector — Function

reduce_object_vector(oVecV, reduceFct; fieldTypes)

Given a vector of mutable structs, compute a function of all numeric fields. Immutable structs are skipped.

Example use case: Compute the mean and standard deviation of simulated model statistics.

reduceFct takes a vector of numeric inputs and returns a single numeric output that is of compatible size and type so it can be written back into an instance of the objects contained in oVecV.

To handle Nan or missing, make sure that reduceFct appropriately handles or skips them.

Optional input fieldTypes indicates which non-numeric field types to use as well.

Limitation: Because the output object is of the same type as the input objects in oVecV, it is not possible to apply reduceFct that returns a different type from its inputs. For example, mean of integers does not work.

Change: exclude immutable fields, but keep scalars (which are immutable but can be changed!) +++

Traversing objects

apply_fct_to_object traverses an object and applies a function to each property. The results are stored in ApplyFctResult.

StructLH.apply_fct_to_object — Function

apply_fct_to_object(x, xName, f)

Apply function f to all properties of an object and its children. Stop parsing child objects if continue_parsing(x) == false. Return tuple with

name of object

type of object

return value of f

entries for children

Beware: Some functions, such as sum, have fallback methods that apply to Any. That leads to crashes.

StructLH.ApplyFctResult — Type

Object that holds result from traversing an object and applying a function to each property.

StructLH.obj_name — Function

obj_name(x)

Name of object.

StructLH.obj_type — Function

obj_type(x)

Type of object.

StructLH.fct_value — Function

fct_value(x)

Function value or nothing.

StructLH.children — Function

children(x)

Child objects, excluding those that contain no function values (themselves, or their children).

The structure of an object and its children can be summarized with one call to describe.

StructLH.describe — Function

describe(x)

Returns a description for an object. Nothing is the default.

StructLH.describe_object — Function

describe_object(x)

Describe an object and its children. Recursively calls describe and returns a vector of strings.

The default return value for describe is nothing. Then the object is not displayed. Otherwise, the return is passed through format_describe for formatting.

StructLH.show_description — Function

show_description(io, x)

Write object description to IO.

Converting to Dicts

The main purpose is to make it easier to serialize user defined types.

StructLH.NodeInfo — Type

Struct that contains information about one node in the Dict created by struct2dict.

StructLH.struct2dict — Function

struct2dict(s)

Convert a struct to a nested Dict. Known leaf types, such as strings, arrays become NodeInfo entries in the Dict. structs become their own Dicts.

The purpose is to convert a struct into a Dict of simple "built-in" types that can be serialized. Each leaf entry contains information about the DataType of the leaf, so that the object can be reconstructed. In case the data are saved to a format that does not preserve size information, the size of the object is also saved.

StructLH.dict2struct! — Function

dict2struct!(s, d)

The inverse of struct2dict. Takes a fully initialized MUTABLE struct as input. Copies the values from the Dict into the struct. Mismatches between Dict keys and propertynames(s) are ignored. That means: fields that are missing from the Dict are not initialized. Extra fields in the Dict are ignored.