UtilityFunctionsLH

Utility functions for economic models.

Defines the abstract type UtilityFunction with subtype UtilityOneArg for utility functions that take one argument, such as u(c).

Specific one argument utility functions defined at this point:

UtilityFunctionsLH.CRRA — Type

Utility = c^(1-sigma) / (1-sigma) - 1.0.

Log utility is special case when sigma == 1.0.

A utility function supports the following methods:

UtilityFunctionsLH.utility — Function

utility(uS, cM)

Utility from consumption.

UtilityFunctionsLH.marginal_utility — Function

marginal_utility(uS, cM)

Marginal utility.

UtilityFunctionsLH.euler_deviation — Function

euler_deviation(uS, beta, R, cV)

Euler equation deviation.

UtilityFunctionsLH.cons_growth — Function

cons_growth(uS, betaR)

Consumption growth rate. Constant for CRRA. Otherwise this would have to depend on consumption levels.

UtilityFunctionsLH.pv_consumption — Function

pv_consumption(uS, beta, R, T)

Present value of consumption as multiple of c1. Periods 1...T. First not discounted.

UtilityFunctionsLH.cons_age1 — Function

Consumption at age 1 for given lifetime income.

UtilityFunctionsLH.cons_path — Function

Consumption path.

UtilityFunctionsLH.lifetime_utility — Function

lifetime_utility(uS, beta, R, T, ltIncome)

Lifetime utility, given present value of income.

lifetime_utility(utilV, beta, T)

Lifetime utility for a given utility sequence.

UtilityFunctionsLH.mu_wealth — Function

mu_wealth(uS, beta, R, T, ltIncome)

Marginal utility of initial assets given present value of income.