EconometricsLH

This package contains general purpose code related to regressions and their display.

RegressorInfo

The RegressorInfo struct holds information about one regressor.

Access functions are: regr_name, regr_coeff, and regr_se.

scale_regressor can be used to rescale a regressor. It just multiplies coefficient and standard error by a scalar.

reduce_regr_infos applies a function, such as mean to all coefficients and standard errors in a vector of regression coefficients. This can be used to average regressions.

EconometricsLH.RegressorInfo — Type

RegressorInfo

Holds info about one regressor

Missing docstring.

Missing docstring for regr_name. Check Documenter's build log for details.

Missing docstring.

Missing docstring for regr_coeff. Check Documenter's build log for details.

Missing docstring.

Missing docstring for regr_se. Check Documenter's build log for details.

Missing docstring.

Missing docstring for scale_regressor. Check Documenter's build log for details.

EconometricsLH.reduce_regr_infos — Function

reduce_regr_infos(rV, reduceFct)

Applies reduceFct to vector of regression coefficients and std errors.

Returns a RegressorInfo. Do not validate this. The user may call a function that does not return a valid RegressorInfo but is still useful.

RegressionTable

A container for storing regression results (currently coefficients and their standard errors).

The intended use cases are:

One receives regression results generated outside of Julia and needs to store them, for example as target moments in an indirect inference problem.

One would like to store the results from regressions run with GLM.jl without storing the associated data, fitted values, etc.

Regression tables are modified with add_regressor, drop_regressor!, change_regressor!, and rename_regressor.

Regressors can be retrieved by name using get_regressor, where has_regressor checks whether a regressor exists. One can also directly retrieve regression coefficients using get_coefficient or standard errors with get_std_error.

A named collection of coefficients and standard errors is retrieved using get_coeff_se_multiple

EconometricsLH.RegressionTable — Type

RegressionTable

Holds regression coefficients and std errors. The order is determined at construction.

Retrieval

EconometricsLH.get_regressor — Function

get_regressor(rt, name)

Retrieve a RegressorInfo object.

Missing docstring.

Missing docstring for has_regressor. Check Documenter's build log for details.

EconometricsLH.get_coefficient — Function

get_coefficient(rt, name)

Return coefficient Coefficient name may be Symbol or String

Missing docstring.

Missing docstring for get_std_error. Check Documenter's build log for details.

EconometricsLH.get_coeff_se — Function

get_coeff_se(rt, name)

Return coefficient and std error as tuple.

EconometricsLH.get_coeff_se_multiple — Function

get_coeff_se_multiple(rt, names)

Return multiple coefficients and std errors.

EconometricsLH.get_all_coeff_se — Function

get_all_coeff_se(rt)

Return all coefficients and std errors.

Modification

EconometricsLH.add_regressor — Function

add_regressor(rt, name, coeff, se)

Add a regressor that does not already exist.

EconometricsLH.drop_regressor! — Function

drop_regressor!(rt, name)

Drop regressors by name.

EconometricsLH.rename_regressor — Function

rename_regressor(rt, oldName, newName)

Rename a regressor.

EconometricsLH.change_regressor! — Function

change_regressor!(rt, ri)

Change regressor.

Regression tests

The purpose is to test whether simulated values match theoretical expected values.

Example: Solve and simulate a model. Check the theoretical value function by comparing it with simulated values.

This is done by regressing simulated values on expected values. This is basically a regression with measurement error in the dependent variable. It should produce a 45 degree line as the OLS regression line. regression_test tests that.

Missing docstring.

Missing docstring for regression_test. Check Documenter's build log for details.