FilesLH

Comparing and Manipulating Directories

FilesLH.clear_directory — Function

clear_directory(tgDir; askFirst)

Clear a directory and its sub-directories

FilesLH.is_dir_empty — Function

is_dir_empty(d)

Is a directory empty?

FilesLH.make_dir — Function

make_dir(d)

Make a directory for a full path.

test this +++++

FilesLH.find_common_base_dir — Function

find_common_base_dir(pathV)

Find common base directory for a set of absolute paths.

FilesLH.files_in_dir — Function

files_in_dir(d)

Make a list of all files in a given directory. Recurses sub-directories. Each line is of the form "subdir/file.ext"

FilesLH.files_not_in_dir2 — Function

files_not_in_dir2(dir1, dir2)

Make a list of files that occur in dir1, but not in dir2. Recurses sub-directories. Only compares file names.

FilesLH.dir_diff_report — Function

dir_diff_report(dir1, dir2; io)

Report differences between two directories and their sub-directories. Based on file names only.

Handling File Names

FilesLH.add_extension — Function

add_extension(filePath, newExt)

Add extension to a file, unless it already has one.

FilesLH.change_extension — Function

change_extension(filePath, newExt)

Change or add extension to a path. newExt may or may not include the leading ".".

FilesLH.extensions_equal — Function

extensions_equal(f1, f2)

Are two extensions equal, ignoring leading "."?

FilesLH.has_extension — Function

has_extension(filePath)

Does file have an extension?

has_extension(filePath, fExt)

Does file have a particular extension? Ignoring leading '.'.

FilesLH.filesep — Function

filesep()

File separator for current OS.

FilesLH.paths_equal — Function

paths_equal(f1, f2)

Compare two paths, taking into account that trailing file separators do not count.

File Transfer

FilesLH.deploy_docs — Function

deploy_docs(pkgDir; trialRun, tgBaseDir)

Deploy package docs to my website.

This is to be included in each package's make.jl. The benefit over hosting on github is to work around issues with private repos and private registries that would arise when TravisCI is used.

Example

# This gives the package directory, if called from `docs/make.jl`.
pkgDir = rstrip(normpath(@__DIR__, ".."), '/');
@assert endswith(pkgDir, "FilesLH")
deploy_docs(pkgDir);

Arguments

pkgDir: points to package directory. If pkgDir = /a/b/Foo the docs end up in julia/Foo on my website.

FilesLH.rsync_command — Function

rsync_command(srcDirIn, tgDirIn; trialRun, doDelete, doCompress, verbose, inclSubDirs, progress, excludeV)

Rsync to or from remote machine. Only for directories.

Fails if local or remote paths contain spaces.

Arguments

srcDirIn, tgDirIn :: String Source and target directories. Remote directory must include part required for ssh (e.g. lhendri@longleaf.unc.edu:)

doDelete :: Bool Should orphan files be deleted on target? False by default.

trialRun :: Bool

progress :: Bool Show progress indicator

#ToDo: How to make the interpolation work when there are quotes around the objects?

Missing docstring.

Missing docstring for rsync_dir. Check Documenter's build log for details.

FilesLH.remote_copy — Function

remote_copy(localPath, remotePath, sshStr, upDown; trialRun)

Copy a file between local and remote computer using scp. scp replaces existing files without warning.

Syntax

`scp /path/local lhendri@longleaf.unc.edu:/path/remote`

Arguments

sshStr :: String

e.g. "lhendri@longleaf.unc.edu"

FilesLH.make_remote_dir — Function

make_remote_dir(remoteDir, sshStr; trialRun, createParents)

Make a directory on the remote, optionally creating all parents as well. Creates a command of the form ssh lhendri@longleaf.unc.edu 'mkdir -p ~/home/abc'. Shows on-screen error if the directory already exists.

Generates on-screen output from ssh login.

Note that other shell commands can be run with the same syntax.

See https://unix.stackexchange.com/questions/8612/programmatically-creating-a-remote-directory-using-ssh

Arguments

sshStr: e.g. "lhendri@longleaf.unc.edu

FilesLH.git_upload_dir — Function

git_upload_dir(repoDir; trialRun)

Upload a repo to git.