ModelStatisticsLH

A package that offers containers for storing model generated statistics by group.

VarInfo

Contains information about a variable:

name

description

data type and size

optional additional meta info stored as key value pairs

ModelStatisticsLH.VarInfo — Type

Holds meta information about one variable. options holds predefined options and user defined options. Predefined options are:

size (required)

lb, ub

lbs, ubs

dimSums

grandSum

These can be used to check a data matrix.

ModelStatisticsLH.get_options — Function

Get multiple options. nothing if not found.

Missing docstring.

Missing docstring for get_options. Check Documenter's build log for details.

Missing docstring.

Missing docstring for has_option. Check Documenter's build log for details.

ModelStatisticsLH.var_name — Function

Returns name of the variable.

Missing docstring.

Missing docstring for description. Check Documenter's build log for details.

ModelStatisticsLH.bounds — Function

Returns the bounds of the variable. These are nothing if not set.

Missing docstring.

Missing docstring for init_values. Check Documenter's build log for details.

ModelStatisticsLH.check_var — Function

check_var(vi, x; silent)

Check an array against the constraints implied by a ArrayInfo.

Arguments

silent: governs whether @warn messages are shown.

ModelStats

Holds statistics and meta information.

Values can be accessed and set using dot notation:

g.x == get_values(g, :x);
z = g.x[1:2, 3, 4];
g.x = [1,2,3];
g.x[1:2, 3:5] = rand(2,3);
g.x[1:2, 3:5] .= 0.5;

ModelStatisticsLH.ModelStats — Type

Container that holds model statistics (arrays or scalars) and meta information. For each variable, a VarInfo object holds info needed to display and check the data.

ModelStatisticsLH.get_meta — Function

get_meta(g, mName)

Retrieve meta info (not for a variable).

ModelStatisticsLH.has_variable — Function

has_variable(g, vName)

Do GroupStats contain variable vName?

ModelStatisticsLH.n_vars — Function

n_vars(g)

Number of variables.

ModelStatisticsLH.var_names — Function

var_names(g)

Return all variable names.

ModelStatisticsLH.var_meta — Function

var_meta(g, vName)

Return meta info for a variable. Errors if not found.

Missing docstring.

Missing docstring for validate_variable. Check Documenter's build log for details.

ModelStatisticsLH.validate_stats — Function

validate_stats(g; silent)

Check that

all data are inside bounds

Missing docstring.

Missing docstring for data_table. Check Documenter's build log for details.

ModelStatisticsLH.add_variable! — Function

add_variable!(g, vInfo, newValues)

Add a variable.

Arguments

vInfo: Can be a VarInfo or a Tuple of values that can be fed into the VarInfo constructor.

Example

add_variable!(g, (:vName, 0.0, 1.0), [1,2,3]);

ModelStatisticsLH.delete_variable! — Function

delete_variable!(g, vName)

Delete a variable. Errors if it does not exist.

ModelStatisticsLH.get_values — Function

get_values(g, vName)

Return values for a variable. Errors if not found.

ModelStatisticsLH.set_values! — Function

set_values!(g, vName, newValues; validate, silent)

Set values for a variable. Errors if variable does not exist. Optional: validates values against meta info.

Example

set_values!(g, :x, [1,2,3]);
# This can also be called with dot notation
g.x = [1,2,3];